# doc/yaml/api.yaml
openapi: 3.0.3
info:
  title: "Squealer Backend"
  description: "Backend API for all Squealer Products"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /users/{handle}/messages/:
    get:
      tags:
        - Message
        - User
      description: Get all the messages of a specific user
      operationId: getUserMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: page
          in: query
          description: Paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
        - name: popular
          in: query
          description: Get only pupular messages, value determines the timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages, value determines the timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages, value determines the timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name and if it starts with
            a \# it will be interpreted as a keyword channel. Otherwise this field will be ignored.
          schema:
            type: string 
            example: ['@feltri', '§canalebello67']
        - name: answering
          in: query 
          description: Will only return messages answering to the given message's ignored.
          schema: 
            type: string 
            example: 653a84466bd663444681acfc
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
    post:
      tags:
        - Message
      description: Post a new Message. This operation will also adjust the poster's character quotas,
        if there are not enough characters it will fail with error 418.
      operationId: postUserMessage
      security:
        - adminAuth: []
        - userAuth: []
      requestBody:
        description: message to be posted
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                  properties:
                    text: 
                      type: string
                      example: This is an example squeal that contains very interesting info.
                    image:
                      type: string
                      format: binary
                      description: An image
                dest:
                  type: array
                  items:
                    type: string
                    description: The handle for single individuals, the channel name for channels
                  example: ['@gianni', '@alberto', '§canalebello']
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Message posted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation.
        '409':
          description: Invalid handle.
        '418': 
          description: User did not have enough characters left.
          
    delete:
      tags:
        - Message
      description: Delete all messages from the user with the given handle
      operationId: deleteUserMessages
      security:
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Messages Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /users/{handle}/messages/{id}:
  
    delete:
      tags:
        - User
        - Message
      description: Delete the message with the given id from the user with the given handle.
      operationId: deleteMessage
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int128
      responses:
        '200':
          description: Message Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
    post:
      tags:
        - User
        - Message
      description: Change message data. Currently only the reactions can be changed by an admin.
      operationId: postMessage
      security:
        - adminAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int64
      requestBody:
        required: True
        description: The data to be changed
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/MessageImpressions'
                - type: object 
                  properties:
                    text: 
                      type: string 
                      description: Edited Message's text,
      responses:
        '200':
          description: Message Modified
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /messages/up/{id}:
  
    delete:
      tags:
        - User
        - Message
      description: Delete a positive reaction.
      operationId: deletePositiveReaction
      security:
        - userAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int128
      responses:
        '200':
          description: Reaction Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid id or No reaction to delete.
    post:
      tags:
        - User
        - Message
      description: Add a positive reaction to a message.
      operationId: postMessage
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int64
      requestBody:
        required: True
        description: The data to be changed
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/MessageImpressions'
                - type: object 
                  properties:
                    text: 
                      type: string 
                      description: Edited Message's text,
      responses:
        '200':
          description: Message Modified
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /messages/down/{id}:
  
    delete:
      tags:
        - User
        - Message
      description: Delete a negative reaction.
      operationId: deleteNegativeReaction
      security:
        - userAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int128
      responses:
        '200':
          description: Reaction Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid id or No reaction to delete.
    post:
      tags:
        - User
        - Message
      description: Add a negative reaction to a message.
      operationId: addNegetiveReaction
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int64
      requestBody:
        required: True
        description: The data to be changed
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/MessageImpressions'
                - type: object 
                  properties:
                    text: 
                      type: string 
                      description: Edited Message's text,
      responses:
        '200':
          description: Message Modified
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
...
  /messages/: 
    get:
      tags:
        - Message
      description: Get all messages. Filtering is strongly advised.
      operationId: getMessages
      parameters:
        - name: page
          in: query
          description: To paginate results. A non-positive number disables pagination.
          schema:
            type: integer
            default: 1
        - name: results_per_page
          in: query
          description: Number of results to return in each page.
          schema:
            type: integer
            default: 100
            minimum: 1
        - name: popular
          in: query
          description: Get only pupular messages within the given timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages within the given timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages within the given timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name. Otherwise this field will be ignored.
            Will only returned messages addressed to ALL valid destinations provided. If one of the given values 
            has no matching user/channel, it will be ignored.
          schema:
            type: string 
            example: ['@feltri', '§canalebello67']
        - name: keywords
          in: query 
          description: Array of the keywords to look for. For each keyword, leading and trailing withe spaces
            will be removed and a \# will be added at the start if not already present. This will filter 
            for messages containing all the keywords. Matching is case-insensitive.
          schema:
            type: array 
            items:
              type: string 
            example: ['#keyword1', 'keyword2']
        - name: mentions
          in: query 
          description: Array of the mentions to look for. For each mention, leading and trailing withe spaces
            will be removed and a @ will be added at the start if not already present. This will filter 
            for messages containing all the mentions. Matching is case-insensitive.
          schema:
            type: array 
            items:
              type: string 
            example: ['@fvPro', 'user12345']
        - name: text 
          in: query
          description: Will only return messages containing the text. Matching is case-insensitive.
          schema: 
            type: string 
            example: lorem ipsum
        - name: answering
          in: query 
          description: Will only return messages answering to the given message's ignored.
          schema: 
            type: string 
            example: 653a84466bd663444681acfc
        - name: publicMessage 
          in: query 
          description: Only return public/private messages. 
          schema:
            type: boolean 
            example: true
        - name: official 
          in: query 
          description: Only return officiale messages. 
          schema:
            type: boolean 
            example: true
      responses:
        200: 
          description: Ok 
          content:
            application/json:
              schema: 
                type: array 
                items:
                  $ref: '#/components/schemas/Message'
  /messages/user/{handle}: 
    get:
      tags:
        - Message
        - User
      description: Get all the messages of a specific user
      operationId: getUserMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: integer
        - name: popular
          in: query
          description: Get only pupular messages, value determines the timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages, value determines the timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages, value determines the timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name and if it starts with
            a \# it will be interpreted as a keyword channel. Otherwise this field will be ignored.
          schema:
            type: string 
            example: ['#sanremo1312', '@feltri', '§canalebello67']
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /messages/user/{handle}/up: 
    get:
      tags:
        - Message
        - User
      description: Get all the messages a specific user liked.
      operationId: getLikedMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: number
            example: 4
            default: 1
        - name: results_per_page
          in: query 
          description: The number of results to be returned 
          schema: 
            type: number
            minimum: 1
            example: 50
            default: 100
      responses:
        '200':
          description: Operation Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
...
  /messages/channel/{name}/:
    get:
      description: Get messages addressed to a given channel. Must have reading permission.
      tags:
          - Channel
      operationId: getChannelMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          required: True
          description: The channel's name.
          schema:
            type: string
            example: awesomeChannel9000
        - name: page
          in: query
          description: Paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
      responses:
        '200': 
          description: Request Successful.
          content:
            application/json:
              schema:
                type: array 
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
    delete:
      description: Remove channel from the dest of all messages.
      tags:
          - Channel
      operationId: deleteChannelMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: name
          in: path
          required: True
          description: The channel's name.
          schema:
            type: string
            example: awesomeChannel9000
      responses:
        '200': 
          description: Request Successful.
        '401':
          description: Unauthorized operation.
        '409':
          description: No channel with the given name.
...
  /public/messages:
    get:
      tags:
        - Message 
        - Public
      description: Return messages available to non-logged in users.
      operationId: publicGetMessages
      parameters:
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
        - name: results_per_page
          in: query
          description: Number of results to return in each page.
          schema:
            type: integer
            default: 100
            minimum: 1
        - name: popular
          in: query
          description: Get only pupular messages within the given timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages within the given timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages within the given timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name. Otherwise this field will be ignored.
            Will only returned messages addressed to ALL valid destinations provided. If one of the given values 
            has no matching user/channel, it will be ignored.
          schema:
            type: string 
            example: ['@feltri', '§canalebello67']
        - name: keywords
          in: query 
          description: Array of the keywords to look for. For each keyword, leading and trailing withe spaces
            will be removed and a \# will be added at the start if not already present. This will filter 
            for messages containing all the keywords. Matching is case-insensitive.
          schema:
            type: array 
            items:
              type: string 
            example: ['#keyword1', 'keyword2']
        - name: mentions
          in: query 
          description: Array of the mentions to look for. For each mention, leading and trailing withe spaces
            will be removed and a @ will be added at the start if not already present. This will filter 
            for messages containing all the mentions. Matching is case-insensitive.
          schema:
            type: array 
            items:
              type: string 
            example: ['@fvPro', 'user12345']
        - name: text 
          in: query
          description: Will only return messages containing the text. Matching is case-insensitive.
          schema: 
            type: string 
            example: lorem ipsum
      responses:
        200: 
          description: Ok 
          content:
            application/json:
              schema: 
                type: array 
                items:
                  $ref: '#/components/schemas/Message'
...
components:
  schemas:
...
    Message:
      allOf:
        - $ref: '#/components/schemas/MessageInfo'
        - $ref: '#/components/schemas/MessageImpressions'
        - type: object
          required: [id, meta]
          properties:
            id:
              type: integer
              format: int64
              description: The message's numerical id
            meta:
              $ref: '#/components/schemas/MessageMeta'
    MessageImpressions:
      type: object
      required: [viewed, reactions]
      properties:
        viewed:
              type: integer
              description: Number of time the message has been viewed
              minimum: 0
        reactions:
          type: object
          required: [positive, negative]
          description: Bookkeeping data on reactions
          properties:
            positive:
              type: object
              description: Numbers of positive reactions divided by types
              properties:
                agree:
                  type: integer
                  minimum: 0
                strongAgree: 
                  type: integer
                  minimum: 0
            negative:
              type: object
              description: Numbers of negative reactions divided by types
              properties:
                disagree:
                  type: integer
                  minimum: 0
                strongDisagree: 
                  type: integer
                  minimum: 0
    MessageInfo:
      type: object
      required: [content, posted, author, dest]
      properties:
        content:
          type: object
          properties:
            text: 
              type: string
              example: This is an example squeal that contains very interesting info.
            image:
              type: string
              format: binary
              description: An image
        posted:
          type: string
          description: Message's timestamp
          format: date-time
          example: "2021-01-30T08:30:00Z"
        author:
          type: string
          description: The author's handle
          example: pieraldo196
        dest:
          type: array
          items:
            type: string
            description: The handle for single individuals, the channel name for channels
          example: ['@gianni', '@alberto', '§canalebello']
    MessageMeta:
      type: object 
      properties:
        created:
          type: string 
          format: date 
          description: Message creation date 
        lastModified:
          type: string 
          format: date 
          description: Last time the message was modified
        geo:
          $ref: '#/components/schemas/GeoSchema'
        impressions:
          type: number 
          description: Number of times the message was viewed.
          minimum: 0
    GeoSchema:
      type: object 
      properties:
        type:
          type: string
          enum: 
            - Point
          default: Point 
        coordinates:
          type: array 
          items: 
            type: number
          maxLength: 2
          minLength: 2
          example: [45.89, 42.069]
...
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires Requires a JWT token generated with user credentials. Even with a correctly formed JWT, a user can only alter their own account.
    proAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: Requires a JWT token generated with pro user credentials.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires a JWT token generated with admin credentials.      